cleaner structure?
