+++ apps/docs/src/examples/shaders/useFboScope/scene/Duck.svelte
