+++ apps/docs/src/examples/shaders/useFboScope/scene/Island.svelte
