+++ apps/docs/src/examples/shaders/useFboScope/scene/Scope.svelte
