+++ apps/docs/src/examples/shaders/useFboScope/scene/Water.svelte
